Included changes to build with build.bat, Also includes windows naming conflicts with some raylib functions. Its messy but was easiest solution

+++ src/external/sinfl.h
@@ -170,7 +170,7 @@ extern int zsinflate(void *out, int cap, const void *in, int size);
 #endif
 
 static int
-sinfl_bsr(unsigned n) {
+sinfl_bsr(unsigned long n) {
 #if defined(_MSC_VER) && !defined(__clang__)
   _BitScanReverse(&n, n);
   return n;
